fix: encapsulate screenshot capture service and localize includeScreenshot state
