+++ shell/src/app/chat/chat-coordinator/chat-coordinator.spec.ts
@@ -51,7 +51,6 @@ class MockAppConfigProvider {
   readonly envMode = signal(EnvMode.STANDALONE);
   readonly authType = signal(AuthType.THIRD_PARTY);
   readonly themePreference = signal<ThemePreference>(ThemePreference.LIGHT);
-  readonly includeScreenshot = signal<boolean>(true);
   setRendererUrl = vi.fn((url: string) => {
     this.rendererUrl.set(url);
   });
@@ -61,9 +60,6 @@ class MockAppConfigProvider {
   setForcedAuthMode = vi.fn();
   setThemePreference = vi.fn((theme: ThemePreference) => {
     this.themePreference.set(theme);
-  });
-  setIncludeScreenshot = vi.fn((include: boolean) => {
-    this.includeScreenshot.set(include);
   });
   flushConfig = vi.fn();
 }
