+++ shell/src/app/chat/llm-client/llm-client.spec.ts
@@ -107,7 +107,6 @@ class MockAppConfigProvider extends AppConfigProvider {
   override readonly rendererUrl = signal<string>('http://mock-renderer.com');
   override readonly geminiApiKey = signal<string>('mock-key-123');
   override readonly themePreference = signal<ThemePreference>(ThemePreference.LIGHT);
-  override readonly includeScreenshot = signal<boolean>(true);
 
   override readonly isApiKeyProvidedByConfig = signal<boolean>(false);
 
@@ -120,9 +119,6 @@ class MockAppConfigProvider extends AppConfigProvider {
   override setForcedAuthMode(mode: AuthType): void {}
   override setThemePreference(theme: ThemePreference): void {
     this.themePreference.set(theme);
-  }
-  override setIncludeScreenshot(include: boolean): void {
-    this.includeScreenshot.set(include);
   }
   override flushConfig(): void {}
 }
