+++ shell/src/app/shell/composer-workspace/composer-workspace.spec.ts
@@ -97,12 +97,8 @@ class MockAppConfigProvider {
   readonly rendererUrl = signal('http://localhost:4200/renderer');
   readonly geminiApiKey = signal('');
   readonly themePreference = signal<ThemePreference>(ThemePreference.LIGHT);
-  readonly includeScreenshot = signal<boolean>(true);
   setThemePreference = vi.fn((theme: ThemePreference) => {
     this.themePreference.set(theme);
-  });
-  setIncludeScreenshot = vi.fn((include: boolean) => {
-    this.includeScreenshot.set(include);
   });
 }
