+++ shell/src/app/shell/screenshot/screenshot-capture.service.ts
@@ -16,24 +16,70 @@
 
 import {Injectable} from '@angular/core';
 
+declare global {
+  class RestrictionTarget {
+    static fromElement(element: Element): Promise<RestrictionTarget>;
+  }
+
+  interface MediaStreamTrack {
+    restrictTo(target: RestrictionTarget | null): Promise<void>;
+  }
+}
+
 /**
  * Service providing screenshot capture capabilities using browser media APIs.
  */
 @Injectable({providedIn: 'root'})
 export class ScreenshotCaptureService {
   /**
-   * Captures a screenshot of the current tab.
+   * Captures a screenshot of the current tab, optionally restricted to a target element.
+   * @param targetElement Optional DOM element to restrict the screenshot to.
    * @returns A base64-encoded PNG image string.
    */
-  async captureScreenshot(): Promise<string> {
+  async captureScreenshot(targetElement?: Element | null): Promise<string> {
+    if (targetElement === null) {
+      throw new Error('No active target element found to capture screenshot.');
+    }
     if (!navigator?.mediaDevices?.getDisplayMedia) {
       throw new Error('Screen capture is not supported in this browser.');
     }
-    const stream = await navigator.mediaDevices.getDisplayMedia({
-      video: {displaySurface: 'browser'},
-      audio: false,
-    });
+    let stream: MediaStream | null = null;
     try {
+      // prettier-ignore
+      stream = await navigator.mediaDevices.getDisplayMedia({
+        video: {displaySurface: 'browser'},
+        audio: false,
+        // @ts-expect-error - preferCurrentTab is a recent/experimental API not yet in TS types
+        'preferCurrentTab': true,
+      });
+
+      const [track] = stream.getVideoTracks();
+      if (!track) {
+        throw new Error('No video track found in media stream.');
+      }
+      if (targetElement) {
+        const restrictionTargetClass = (globalThis as Record<string, unknown>)[
+          'RestrictionTarget'
+        ] as {fromElement?: (element: Element) => Promise<RestrictionTarget>} | undefined;
+
+        if (
+          typeof restrictionTargetClass?.fromElement === 'function' &&
+          typeof track?.restrictTo === 'function'
+        ) {
+          try {
+            const target = await restrictionTargetClass.fromElement(targetElement);
+            await track.restrictTo(target);
+          } catch (restrictionError) {
+            console.warn(
+              'Failed to restrict video track to element, falling back to full tab capture:',
+              restrictionError,
+            );
+          }
+        } else {
+          console.warn('RestrictionTarget API not supported, capturing full tab.');
+        }
+      }
+
       const video = document.createElement('video');
       video.muted = true;
       video.playsInline = true;
@@ -44,6 +90,10 @@ export class ScreenshotCaptureService {
         };
         video.onerror = reject;
       });
+
+      // Give the browser a tiny buffer to apply the restriction crop visually
+      await new Promise(resolve => setTimeout(resolve, 150));
+
       const canvas = document.createElement('canvas');
       canvas.width = video.videoWidth;
       canvas.height = video.videoHeight;
@@ -51,6 +101,9 @@ export class ScreenshotCaptureService {
       if (!ctx) throw new Error('Failed to create canvas context for screenshot');
       ctx.drawImage(video, 0, 0, canvas.width, canvas.height);
       return canvas.toDataURL('image/png');
+    } catch (error) {
+      console.warn('Capture canceled or failed:', error);
+      throw error;
     } finally {
       if (stream) stream.getTracks().forEach(track => track.stop());
     }
